+++ src/DETHRACE/common/depth.c
@@ -469,7 +469,7 @@ void ExternalSky(br_pixelmap* pRender_buffer, br_pixelmap* pDepth_buffer, br_act
     tan_half_hori_fov = tan_half_fov * camera->aspect;
 
     vert_sky = BrRadianToAngle(atan2(pCamera_to_world->m[2][0], pCamera_to_world->m[2][2]));
-    hori_sky = BrRadianToAngle(atan2(col_map->width * tan_half_hori_fov / (double)pRender_buffer->width, 1));
+    hori_sky = BrRadianToAngle(atan2(col_map->width * tan_half_hori_fov / (float)pRender_buffer->width, 1));
 
     tan_half_hori_sky = -BrFixedToFloat(vert_sky) / BrFixedToFloat(BR_ANGLE_DEG(360) / (int)(1.0f / BrFixedToFloat(2 * hori_sky) + 0.5f));
